requirements.txt
chromadb
whoosh
minio
networkx
python-dotenv
pyyaml
sentence-transformers
pydantic
fastapi
langchain-community
pypdf
openai
